I think I only changed the floatX decleration in tensor_utlis from float to int. It seems that the DataTypeToEnum supports int but the funtionality of the code itself can't handle int so we'll have to fix it I suppose.

+++ nanolib2.0/tensor/tensor_util.hpp
@@ -4,7 +4,7 @@
 #include "absl/container/inlined_vector.h"
 #include "tensor_types.hpp"
 
-using floatX = float;
+using floatX = int;
 
 // Raw pointer -> Flat
 template <typename T>
